# [enhancement]Optimize GeoFunctions for const columns