## Proposed changes

Issue Number: #34131 

<!--Describe your changes.-->
1000w row data
st_point()
```
before avg: 4.36s
now avg:4.338s
```
st_azimuth()
```
before avg:4.215s
now avg:3.718s
```

st_contains()
```
before avg:4.22s
now avg:3.8275s
```
test sql
```
set parallel_pipeline_task_num=1;
pager cat > /dev/null

select st_point(1.0,a2) from const_double;
select st_azimuth(st_point(1.0,0.0),a2) from geo;
select st_contains(ST_Polygon("POLYGON ((0 0, 10 0, 10 10, 0 10, 0 0))"), a2) from geo;
```
## Further comments

If this is a relatively large or complex change, kick off the discussion at [dev@doris.apache.org](mailto:dev@doris.apache.org) by explaining why you chose the solution you did and what alternatives you considered, etc...