@koarz 2024-05-05
run buildall