@koarz 2024-05-06
run buildall