@koarz 2024-05-07
run buildall